Made changes to allow for consecutive database scaffolding
To enable consecutive database scaffolding, I moved all modifications to the `DbContext` into a partial `DbContext` class. This ensures that any changes are preserved in the partial class whenever the database is scaffolded again. *Note:* Scaffolding *will* produce compile-time errors, which can be resolved by removing duplicate methods from the automatically generated `DbContext`.
