+++ BumboApp/Bumbo.Data/Bumbo.Data.csproj
@@ -11,4 +11,10 @@
       <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="8.0.10" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Compile Update="Context\BumboDbContext.Partial.cs">
+        <DependentUpon>BumboDbContext.cs</DependentUpon>
+      </Compile>
+    </ItemGroup>
+
 </Project>
