+++ BumboApp/Bumbo.Data/Context/BumboDbContext.cs
@@ -2,24 +2,11 @@
 using System.Collections.Generic;
 using Bumbo.Data.Models;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
 
 namespace Bumbo.Data.Context;
 
 public partial class BumboDbContext : DbContext
 {
-    private readonly IConfiguration _configuration;
-    
-    public BumboDbContext(IConfiguration configuration)
-    {
-        _configuration = configuration;
-    }
-
-    public BumboDbContext(DbContextOptions<BumboDbContext> options, IConfiguration configuration)
-        : base(options)
-    {
-        _configuration = configuration;
-    }
 
     public virtual DbSet<Branch> Branches { get; set; }
 
@@ -42,15 +29,6 @@ public partial class BumboDbContext : DbContext
     public virtual DbSet<SupermarketActivity> SupermarketActivities { get; set; }
 
     public virtual DbSet<Weekday> Weekdays { get; set; }
-
-    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-    {
-        if (!optionsBuilder.IsConfigured)
-        {
-            // optionsBuilder.UseSqlServer(_configuration.GetConnectionString("Bumbo"));
-            optionsBuilder.UseSqlServer("Server = .; Database = Bumbo; Integrated Security = True; TrustServerCertificate = True;");
-        }
-    }
 
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
